+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Changed
 - Update commands to better reflect the protocol:
   - `watch()` returns integer of number of watched tubes.
+  - `ignore()` throws `CommandException` if trying to ignore the final tube.
 - Status is a required argument for the `tube:peek` CLI command.
 - **BC break**: Constructor for `Connection` no longer needs a `Socket`.
   Pass the same parameters directly to `Connection`.